done delete

+++ Jewelry Shop_Back-End/controllers/orderController.js
@@ -4,21 +4,26 @@ import {orderRepository} from "../repositories/indexRepository.js"
 const createOrder = async (req, res) => {
     try {
       const cartToken = req.params.cartToken;
-      const userId = req.body.userId;
+      const userId = req.body.user_id;
       const order_date = req.body.order_date;
       const total_amount = req.body.total_amount;
       const orderStatus = req.body.orderStatus;
   
-      const cart = await cartRepository.getCartByUserId(userId);
-  
-      if (!cart) {
-        return res.status(404).json({ message: 'Cart not found' });
+      const cart = await cartRepository.getCartByToken(cartToken);
+      
+      if (!userId) {
+        return res.status(HttpStatusCode.UNAUTHORIZED).json({ message: 'You have to login to buy' });
       }
-  
+      if (!cart) {
+        return res.status(HttpStatusCode.NOT_FOUND).json({ message: 'Cart not found' });
+      }
+      if (cart.user_id !== userId) {
+        return res.status(HttpStatusCode.UNAUTHORIZED).json({ message: 'Cart does not belong to the user' });
+      }
       const order = await orderRepository.createOrder(userId, order_date, total_amount, orderStatus);
   
-      for (const product of cart.products) {
-        await orderRepository.createOrderDetail(order._id, product.productId, product.quantity, product.size, product.color, product.material);
+      for (const product of cart.productList) {
+        await orderRepository.createOrderDetail(order._id, product.product_id, product.quantity, product.size, product.color, product.material);
       }
   
       await cartRepository.removePurchasedProducts(cart._id);
